requirements.txt
python-dotenv
telethon
requests
beautifulsoup4